requirements.txt
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.0
entrypoints==0.3
executing==0.8.2
fastapi==0.73.0
fastapi-health==0.4.0
h11==0.13.0
idna==3.3
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyter-server==1.13.4
jupyterlab==3.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.7
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1643009225447/work
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.2
pyarrow==6.0.1
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pymongo==4.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
pyspark==3.2.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
stack-data==0.1.4
starlette==0.17.1
terminado==0.12.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.17.0.post1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
